+++ architect_py/internal_utils/no_pandas.py
@@ -1,3 +1,0 @@
-from typing import Never
-
-DataFrame = Never